Services/django/requirements.txt
django==6.0
django-extensions==4.1
djangorestframework==3.16.1
djoser==2.3.3
uvicorn[standard]==0.38.0
granian==2.6.0
whitenoise==6.11.0
psycopg[binary]==3.3.2
django-cors-headers==4.9.0